backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
marshmallow==3.20.1
web3==6.15.1
eth-account==0.10.0